misc cleanup

+++ src/store.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: store.cc,v 1.132 1996/10/17 11:14:50 wessels Exp $
+ * $Id: store.cc,v 1.133 1996/10/18 20:36:26 wessels Exp $
  *
  * DEBUG: section 20    Storeage Manager
  * AUTHOR: Harvest Derived
@@ -256,7 +256,7 @@ static int store_pages_high = 0;
 static int store_pages_low = 0;
 
 /* current file name, swap file, use number as a filename */
-static unsigned long swapfileno = 0;
+static int swapfileno = 0;
 static int store_swap_size = 0;	/* kilobytes !! */
 static unsigned long store_swap_high = 0;
 static unsigned long store_swap_low = 0;
